Add tests for updating groupSize / groupCount together with template
This adds tests for using the client to update the groupSize as well as
the groupCount when changing the template of a competition. In addition it
adds the mentioned variables to the documentation of methods in the
Pwned_Client class, as well as method documentation for the tests of the
client.