Major update to the client, adds proper support for bundles, leagues, starting support for ladders and rankings.
See tests for usage.

A bundle is defined as a grouping of different competitions, so that
you're able to keep track of competitions belonging together - such as
different leagues as divisions in a series, etc.

There is no limitation on the different types of competitions that can
be added to a bundle. They're returned in the order created. Add
bundleId that refers to a bundle to your request when creating
a competition.

This also adds support for creating several leagues in one statement.
Include the key leagueCount when creating a league to get several
leagues created. It's strongly suggested to use this feature together
with bundleId to get all the leagues grouped into a bundle as well.