Add support for retiring a player / team from a league.
This patch adds support for retiring an active player from a league. This
will remove all past, present and future matches for the player / team,
and tag the signup as retired. All future standings will have the
player at the bottom of the ranking, and the signup will have the
"retired" key set to true.

This currently only makes sense for standard leagues and not for
championships, and is not handled in any way for tournaments or
other competition types.